Homework57

+++ src/main/java/kg/attractor/jobsearchjava27/dao/UserDao.java
@@ -21,7 +21,7 @@ public class UserDao {
     private final JdbcTemplate jdbcTemplate;
     private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
 
-    public Optional<User> findById(int id) {
+    public Optional<User> findById(Long id) {
         String sql = "SELECT * FROM usr WHERE id = ?";
         try {
             User user = jdbcTemplate.queryForObject(sql, new BeanPropertyRowMapper<>(User.class), id);
